Move SecurityAccess to the API module
