+++ Kitodo-API/src/main/java/org/kitodo/security/SecurityAccess.java
@@ -1,0 +1,153 @@
+/*
+ * (c) Kitodo. Key to digital objects e. V. <contact@kitodo.org>
+ *
+ * This file is part of the Kitodo project.
+ *
+ * It is licensed under GNU General Public License version 3 or later.
+ *
+ * For the full copyright and license information, please read the
+ * GPL3-License.txt file that was distributed with this source code.
+ */
+
+package org.kitodo.security;
+
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Objects;
+
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.context.SecurityContextHolder;
+
+/**
+ * If module wants to use own roles, it needs to extend this class. Currently it
+ * is usable only for global roles as client id is not available to modules.
+ * // TODO: find a way for modules to get information about current client id
+ */
+public class SecurityAccess {
+
+    private static final String GLOBAL_IDENTIFIER = "GLOBAL";
+    private static final String CLIENT_IDENTIFIER = "CLIENT";
+
+    private Collection<? extends GrantedAuthority> getAuthoritiesOfCurrentAuthentication() {
+        Authentication authentication = getCurrentAuthentication();
+        if (Objects.nonNull(authentication)) {
+            return authentication.getAuthorities();
+        } else {
+            return new ArrayList<>();
+        }
+    }
+
+    /**
+     * Check if the current user has a specified authority globally.
+     *
+     * @param authorityTitle
+     *            the authority title
+     * @return true if the current user has the specified authority
+     */
+    public boolean hasAuthorityGlobal(String authorityTitle) {
+        return hasAuthority(authorityTitle + "_" + GLOBAL_IDENTIFIER);
+    }
+
+    /**
+     * Check if the current user has a specified authority for a client.
+     *
+     * @param authorityTitle
+     *            the authority title
+     * @param clientId
+     *            id of current session client
+     * @return true if the current user has the specified authority
+     */
+    public boolean hasAuthorityForClient(String authorityTitle, int clientId) {
+        String clientAuthority = authorityTitle + "_" + CLIENT_IDENTIFIER + "_" + clientId;
+        return hasAuthority(clientAuthority);
+    }
+
+    /**
+     * Check if the current user has a specified authority globally or for a client.
+     *
+     * @param authorityTitle
+     *            the authority title
+     * @param clientId
+     *            id of current session client
+     * @return true if the current user has the specified authority
+     */
+    public boolean hasAuthorityGlobalOrForClient(String authorityTitle, int clientId) {
+        return hasAuthorityGlobal(authorityTitle) || hasAuthorityForClient(authorityTitle, clientId);
+    }
+
+    /**
+     * Check if the current user has any of the specified authorities globally.
+     *
+     * @param authorityTitles
+     *            the authority titles separated with commas e.g. "authority1,
+     *            authority2, authority3"
+     * @return true if the current user has any of the specified authorities
+     *         globally
+     */
+    public boolean hasAnyAuthorityGlobal(String authorityTitles) {
+        String[] authorityTitlesArray = getStringArray(authorityTitles);
+        for (String authorityTitle : authorityTitlesArray) {
+            if (hasAuthorityGlobal(authorityTitle)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Check if the current user has a specified authority for a client.
+     *
+     * @param authorityTitles
+     *            the authority title
+     * @param clientId
+     *            id of current session client
+     * @return true if the current user has the specified authority
+     */
+    public boolean hasAnyAuthorityForClient(String authorityTitles, int clientId) {
+        String[] authorityTitlesArray = getStringArray(authorityTitles);
+        for (String authorityTitle : authorityTitlesArray) {
+            if (hasAuthorityForClient(authorityTitle, clientId)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    /**
+     * Check if the current user has any of the specified authorities globally or
+     * for client.
+     *
+     * @param authorityTitles
+     *            the authority titles separated with commas e.g. "authority1,
+     *            authority2, authority3"
+     * @param clientId
+     *            id of current session client
+     * @return true if the current user has any of the specified authorities
+     *         globally or for client
+     */
+    public boolean hasAnyAuthorityGlobalOrForClient(String authorityTitles, int clientId) {
+        return hasAnyAuthorityGlobal(authorityTitles) || hasAnyAuthorityForClient(authorityTitles, clientId);
+    }
+
+    /**
+     * Get Authentication object of current threads security context.
+     *
+     * @return authentication object
+     */
+    protected Authentication getCurrentAuthentication() {
+        return SecurityContextHolder.getContext().getAuthentication();
+    }
+
+    private boolean hasAuthority(String authorityTitle) {
+        SimpleGrantedAuthority authority = new SimpleGrantedAuthority(authorityTitle);
+        Collection<? extends GrantedAuthority> authorities = getAuthoritiesOfCurrentAuthentication();
+        return authorities.contains(authority);
+    }
+
+    private String[] getStringArray(String strings) {
+        strings = strings.replaceAll("\\s+", ""); // remove whitespaces
+        return strings.split(",");
+    }
+}
