+++ pom.xml
@@ -69,6 +69,7 @@
         <maven-jar-plugin.version>3.1.0</maven-jar-plugin.version>
         <mockito.version>2.0.2-beta</mockito.version>
         <properties-maven-plugin.version>1.0.0</properties-maven-plugin.version>
+        <spring.security.version>4.2.3.RELEASE</spring.security.version>
         <flyway-maven-plugin.version>5.1.4</flyway-maven-plugin.version>
     </properties>
